requirements.txt
# 项目依赖
fastapi
# pymupdf
pdfplumber
numpy
openai
opencv_contrib_python
opencv_python
opencv_python_headless
# PaddlePaddle CPU版本（必须在paddlex之前安装）
paddlepaddle==3.0.0rc0 -i https://www.paddlepaddle.org.cn/packages/stable/cpu/
https://paddle-model-ecology.bj.bcebos.com/paddlex/whl/paddlex-3.0.0rc0-py3-none-any.whl
Pillow
python-dotenv
Requests
tqdm
urllib3
uvicorn
python-multipart


# 开发依赖
pre-commit>=3.6.2
black>=24.2.0
isort>=5.13.2
flake8>=7.0.0
flake8-docstrings>=1.7.0
